+++ src/error.rs
@@ -8,7 +8,7 @@ impl<R, E: ToString> ResultLogger<R> for Result<R, E> {
   fn log_ignore(&self) {
     match self {
       Ok(_) => (),
-      Err(err) => log::error!("Error: {}", err.to_string()),
+      Err(err) => log::debug!("Error: {}", err.to_string()),
     }
   }